+++ src/patch/core/ASettings.zig
@@ -6,6 +6,8 @@ const bufPrintZ = std.fmt.bufPrintZ;
 
 const GlobalSt = @import("../appinfo.zig").GLOBAL_STATE;
 const GlobalFn = @import("../appinfo.zig").GLOBAL_FUNCTION;
+
+const workingOwner = @import("Hook.zig").PluginState.workingOwner;
 const coreAllocator = @import("Allocator.zig").allocator;
 
 const HandleMapSOA = @import("../util/handle_map_soa.zig").HandleMapSOA;
@@ -24,11 +26,11 @@ const NullHandle = Handle.getNull();
 
 const DEFAULT_ID = 0xFFFF; // TODO: use ASettings plugin id?
 
-const SentSetting = extern struct {
+pub const ASettingSent = extern struct {
     name: [*:0]u8,
     value: Value,
 
-    const Value = extern union {
+    pub const Value = extern union {
         str: [*:0]u8,
         f: f32,
         u: u32,
@@ -49,7 +51,7 @@ const SentSetting = extern struct {
 };
 
 // TODO: add global st/fn ptrs to fnOnChange def?
-const Setting = struct {
+pub const Setting = struct {
     section: ?struct { generation: u16, index: u16 } = null,
     name: [63:0]u8 = std.mem.zeroes([63:0]u8),
     value: Value = .{ .str = std.mem.zeroes([63:0]u8) },
@@ -57,18 +59,18 @@ const Setting = struct {
     value_saved: Value = .{ .str = std.mem.zeroes([63:0]u8) },
     value_type: Type = .None,
     flags: EnumSet(Flags) = EnumSet(Flags).initEmpty(),
-    fnOnChange: ?*const fn (value: SentSetting.Value) callconv(.C) void = null,
+    fnOnChange: ?*const fn (value: ASettingSent.Value) callconv(.C) void = null,
 
-    const Type = enum(u8) { None, Str, F, U, I, B };
+    pub const Type = enum(u8) { None, Str, F, U, I, B };
 
-    const Value = extern union {
+    pub const Value = extern union {
         str: [63:0]u8,
         f: f32,
         u: u32,
         i: i32,
         b: bool,
 
-        pub inline fn set(self: *Value, v: SentSetting.Value, t: Type) !void {
+        pub inline fn set(self: *Value, v: ASettingSent.Value, t: Type) !void {
             switch (t) {
                 .Str => _ = try bufPrintZ(&self.str, "{s}", .{v.str}),
                 .F => self.f = v.f,
@@ -137,7 +139,7 @@ const Setting = struct {
         DefaultValueNotConverted,
     };
 
-    inline fn sent2setting(setting: SentSetting.Value) Value {
+    inline fn sent2setting(setting: ASettingSent.Value) Value {
         _ = setting;
     }
 };
@@ -147,11 +149,11 @@ const Setting = struct {
 // FIXME: section->index may become invalid when handle expires due to swapRemove
 // in handle_map; same problem with Setting->section too
 // need to confirm if this is an issue, and maybe rework how connecting parents works
-const Section = struct {
+pub const Section = struct {
     section: ?struct { generation: u16, index: u16 } = null,
     name: [63:0]u8 = std.mem.zeroes([63:0]u8),
     flags: EnumSet(Flags) = EnumSet(Flags).initEmpty(),
-    fnOnChange: ?*const fn (changed: [*]SentSetting) callconv(.C) void = null, // null-terminated
+    fnOnChange: ?*const fn (changed: [*]ASettingSent) callconv(.C) void = null, // null-terminated
 
     const Flags = enum(u32) {
         HasOwner,
@@ -233,7 +235,7 @@ const ASettings = struct {
         owner: u16,
         section: ?Handle,
         name: [*:0]const u8,
-        fnOnChange: ?*const fn ([*]SentSetting) callconv(.C) void,
+        fnOnChange: ?*const fn ([*]ASettingSent) callconv(.C) void,
     ) !Handle {
         const existing_i = nodeFind(data_sections, section, name);
 
@@ -261,7 +263,7 @@ const ASettings = struct {
     // FIXME: review how the handle map is being manipulated with respect to index
     // to make sure it's not messing with the mapping
     /// release ownership of a section node, and all of the children below it
-    pub fn sectionRelease(
+    pub fn sectionVacate(
         handle: Handle,
     ) void {
         var data: Section = data_sections.get(handle) orelse return;
@@ -270,7 +272,7 @@ const ASettings = struct {
         for (sec_slice_sections, 0..) |*s, i| {
             if (s.* != null and s.*.?.generation == handle.generation and s.*.?.index == handle.index) {
                 const h: Handle = data_sections.handles.items[i];
-                if (h.owner != DEFAULT_ID and h.owner == handle.owner) sectionRelease(h);
+                if (h.owner != DEFAULT_ID and h.owner == handle.owner) sectionVacate(h);
             }
         }
 
@@ -278,7 +280,7 @@ const ASettings = struct {
         for (set_slice_sections, 0..) |*s, i| {
             if (s.* != null and s.*.?.generation == handle.generation and s.*.?.index == handle.index) {
                 const h: Handle = data_settings.handles.items[i];
-                if (h.owner != DEFAULT_ID and h.owner == handle.owner) settingRelease(h);
+                if (h.owner != DEFAULT_ID and h.owner == handle.owner) settingVacate(h);
             }
         }
 
@@ -321,18 +323,20 @@ const ASettings = struct {
 
     // FIXME: review how the handle map is being manipulated with respect to index
     // to make sure it's not messing with the mapping
+    /// take ownership of a setting
+    /// will cause callback to run on the initial value
     pub fn settingOccupy(
         owner: u16,
         section: ?Handle,
         name: [*:0]const u8,
         value_type: Setting.Type,
         value_default: Setting.Value,
-        fnOnChange: ?*const fn (SentSetting.Value) callconv(.C) void,
+        fnOnChange: ?*const fn (ASettingSent.Value) callconv(.C) void,
     ) !Handle {
         std.debug.assert(value_type != .None);
         if (section) |s| {
             if (s.owner != owner) @panic("owner and handle owner must match");
-            if (!data_sections.hasHandle(s)) return NullHandle;
+            if (!data_sections.hasHandle(s)) return error.SectionDoesNotExist;
         }
 
         const existing_i = nodeFind(data_settings, section, name);
@@ -364,7 +368,7 @@ const ASettings = struct {
         data.flags.insert(.DefaultValueIsSet);
         data.fnOnChange = fnOnChange;
         if (fnOnChange) |f|
-            f(SentSetting.Value.fromSetting(&data.value, data.value_type));
+            f(ASettingSent.Value.fromSetting(&data.value, data.value_type));
 
         if (existing_i) |i| {
             data_settings.values.set(i, data);
@@ -378,7 +382,7 @@ const ASettings = struct {
 
     // FIXME: review how the handle map is being manipulated with respect to index
     // to make sure it's not messing with the mapping
-    pub fn settingRelease(
+    pub fn settingVacate(
         handle: Handle,
     ) void {
         var data: Setting = data_settings.get(handle) orelse return;
@@ -389,9 +393,9 @@ const ASettings = struct {
         data.value_default = .{ .str = std.mem.zeroes([63:0]u8) };
         data.flags.remove(.DefaultValueIsSet);
 
-        data.value.type2raw(data.value_type) catch @panic("settingRelease: 'value' invalid");
+        data.value.type2raw(data.value_type) catch @panic("settingVacate: 'value' invalid");
         if (!data.flags.contains(.SavedValueNotConverted))
-            data.value_saved.type2raw(data.value_type) catch @panic("settingRelease: 'value_saved' invalid");
+            data.value_saved.type2raw(data.value_type) catch @panic("settingVacate: 'value_saved' invalid");
         data.flags.remove(.SavedValueNotConverted);
 
         data.value_type = .None;
@@ -402,15 +406,26 @@ const ASettings = struct {
     }
 
     // TODO: ValueUpdated flag?
+    /// trigger setting update with new value
+    /// will cause callback to run
     pub fn settingUpdate(
         handle: Handle,
-        value: SentSetting.Value,
+        value: ASettingSent.Value,
     ) void {
         var data: Setting = data_settings.get(handle) orelse return;
         data.value.set(value, data.value_type) catch return;
         data_settings.values.set(handle.index, data);
         if (data.fnOnChange) |f|
             f(value);
+    }
+
+    pub fn vacateOwner(owner: u16) void {
+        // settings first for better cache use of data_settings processes
+        for (ASettings.data_settings.handles.items) |h|
+            if (h.owner == owner) ASettings.settingVacate(h);
+
+        for (ASettings.data_sections.handles.items) |h|
+            if (h.owner == owner) ASettings.sectionVacate(h);
     }
 
     // TODO: dumping stuff i might need here
@@ -428,16 +443,56 @@ const ASettings = struct {
 // GLOBAL EXPORTS
 
 // TODO: impl
-pub fn ASectionOccupy() callconv(.C) void {}
-pub fn ASectionRelease() callconv(.C) void {}
-pub fn ASettingOccupy() callconv(.C) void {}
-pub fn ASettingRelease() callconv(.C) void {}
-pub fn ASettingUpdate() callconv(.C) void {}
+pub fn ASectionOccupy(
+    section: Handle,
+    name: [*:0]const u8,
+    fnOnChange: ?*const fn ([*]ASettingSent) callconv(.C) void,
+) callconv(.C) Handle {
+    return ASettings.sectionOccupy(
+        workingOwner(),
+        if (section.isNull()) null else section,
+        name,
+        fnOnChange,
+    ) catch NullHandle;
+}
+
+pub fn ASectionVacate(handle: Handle) callconv(.C) void {
+    ASettings.sectionVacate(handle);
+}
+
+pub fn ASettingOccupy(
+    section: Handle,
+    name: [*:0]const u8,
+    value_type: Setting.Type,
+    value_default: Setting.Value,
+    fnOnChange: ?*const fn (ASettingSent.Value) callconv(.C) void,
+) callconv(.C) Handle {
+    return ASettings.settingOccupy(
+        workingOwner(),
+        if (section.isNull()) null else section,
+        name,
+        value_type,
+        value_default,
+        fnOnChange,
+    ) catch NullHandle;
+}
+
+pub fn ASettingVacate(handle: Handle) callconv(.C) void {
+    ASettings.settingVacate(handle);
+}
+
+pub fn ASettingUpdate(handle: Handle, value: ASettingSent.Value) callconv(.C) void {
+    ASettings.settingUpdate(handle, value);
+}
+
+pub fn AVacateAll() callconv(.C) void {
+    ASettings.vacateOwner(workingOwner());
+}
 
 // HOOKS
 
 // FIXME: remove, for testing (or move to commented test block)
-fn updateSet1(value: SentSetting.Value) callconv(.C) void {
+fn updateSet1(value: ASettingSent.Value) callconv(.C) void {
     dbg.ConsoleOut("set1 changed to {d:4.2}\n", .{value.f}) catch {};
 }
 
@@ -450,7 +505,7 @@ pub fn OnInit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
     _ = ASettings.sectionNew(null, "Sec2") catch {};
     _ = ASettings.sectionNew(null, "Sec2") catch {}; // expect: NameTaken error -> skipped
     const sec1 = ASettings.sectionOccupy(0x0000, null, "Sec1", null) catch Handle.getNull();
-    const sec2 = ASettings.sectionOccupy(0x0000, null, "Sec2", null) catch Handle.getNull();
+    const sec2 = ASettings.sectionOccupy(0x0001, null, "Sec2", null) catch Handle.getNull();
 
     _ = ASettings.settingNew(sec1, "Set1", "123.456", false) catch {};
     _ = ASettings.settingNew(sec1, "Set1", "123.456", false) catch {};
@@ -466,12 +521,14 @@ pub fn OnInit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
     _ = ASettings.settingOccupy(0x0000, null, "Set6", .F, .{ .f = 876.543 }, null) catch {}; // export: ignored
 
     ASettings.settingUpdate(occ1, .{ .f = 678.543 }); // expect: changed value
-    ASettings.settingRelease(occ2); // expect: undefined default, etc.
+    ASettings.settingVacate(occ2); // expect: undefined default, etc.
 
     const sec3 = ASettings.sectionOccupy(0x0000, null, "Sec3", null) catch Handle.getNull();
     _ = ASettings.settingNew(sec3, "Set7", "Val7", false) catch {};
     _ = ASettings.settingOccupy(0x0000, sec3, "Set8", .F, .{ .f = 987.654 }, null) catch NullHandle;
-    ASettings.sectionRelease(sec3);
+    ASettings.sectionVacate(sec3);
+
+    ASettings.vacateOwner(0x0000); // expect: everything undefined default, etc.
 }
 
 pub fn OnInitLate(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
@@ -480,7 +537,9 @@ pub fn OnDeinit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
     ASettings.deinit();
 }
 
-//pub fn OnPluginDeinit(_: u16) callconv(.C) void {}
+pub fn OnPluginDeinit(_: u16) callconv(.C) void {
+    ASettings.vacateOwner(workingOwner());
+}
 
 // FIXME: remove, for testing
 // TODO: maybe adapt for debug/testing